refactor(#9): 파일 이동 및 use case 추가
