+++ src/main/java/com/example/leaflog/bc/account/auth/application/service/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.leaflog.bc.account.auth.application.exception;
+package com.example.leaflog.bc.account.auth.application.service.exception;
 
 import com.example.leaflog.bc.sharedkernel.exception.model.CustomException;
 import com.example.leaflog.bc.sharedkernel.exception.model.ErrorCode;
