+++ src/main/java/com/example/leaflog/bc/profile/profile/application/port/in/ProfileQueryUseCase.java
@@ -1,0 +1,7 @@
+package com.example.leaflog.bc.profile.profile.application.port.in;
+
+import com.example.leaflog.bc.profile.profile.application.dto.ProfileReadModel;
+
+public interface ProfileQueryUseCase {
+    ProfileReadModel query();
+}
