+++ src/main/java/com/example/leaflog/bc/profile/profile/presentation/adapter/in/web/ProfileController.java
@@ -1,7 +1,7 @@
-package com.example.leaflog.bc.profile.profile.presentation;
+package com.example.leaflog.bc.profile.profile.presentation.adapter.in.web;
 
-import com.example.leaflog.bc.profile.profile.application.ProfileQueryService;
-import com.example.leaflog.bc.profile.profile.application.query.ProfileReadModel;
+import com.example.leaflog.bc.profile.profile.application.port.in.ProfileQueryUseCase;
+import com.example.leaflog.bc.profile.profile.application.dto.ProfileReadModel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,10 +12,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/profile")
 public class ProfileController {
 
-    private final ProfileQueryService profileQueryService;
+    private final ProfileQueryUseCase ProfileQueryUseCase;
 
     @GetMapping()
     public ProfileReadModel getProfile(){
-        return profileQueryService.query();
+        return ProfileQueryUseCase.query();
     }
 }
